Refactor routes into blueprints
