+++ admin.py
@@ -40,7 +40,7 @@ class MyModelView(ModelView):
 
     def inaccessible_callback(self, name, **kwargs):
         flash("Acesso restrito à administração.", "danger")
-        return redirect(url_for('login'))
+        return redirect(url_for('auth.login'))
 
 # --------------------------------------------------------------------------
 # Dashboard (será a página inicial do painel)
@@ -51,7 +51,7 @@ class AdminDashboard(BaseView):
 
     def inaccessible_callback(self, name, **kwargs):
         flash("Acesso restrito à administração.", "danger")
-        return redirect(url_for('login'))
+        return redirect(url_for('auth.login'))
 
     @expose('/')
     @login_required
