+++ templates/index.html
@@ -13,7 +13,7 @@
                 <div class="col-md-3">
                     <lottie-player src="https://assets5.lottiefiles.com/packages/lf20_ghs9bthn.json"
                                    background="transparent" speed="1" style="height: 100px;" loop autoplay></lottie-player>
-                    <a href="{{ url_for('add_animal') }}" class="btn btn-outline-success rounded-pill shadow-sm px-4 py-2"
+                    <a href="{{ url_for('animals.add_animal') }}" class="btn btn-outline-success rounded-pill shadow-sm px-4 py-2"
                        data-bs-toggle="tooltip" title="Cadastre um novo animal na plataforma.">
                         ➕ Cadastrar novo animal
                     </a>
@@ -70,9 +70,9 @@
             
             <div class="d-flex justify-content-center gap-3 mt-4">
                 
-                <a href="{{ url_for('register') }}" class="btn btn-outline-success rounded-pill px-4 py-2">🧑 Criar Conta</a>
+                <a href="{{ url_for('auth.register') }}" class="btn btn-outline-success rounded-pill px-4 py-2">🧑 Criar Conta</a>
                
-                <a href="{{ url_for('login') }}" class="btn btn-outline-primary rounded-pill px-4 py-2">🔐 Entrar</a>
+                <a href="{{ url_for('auth.login') }}" class="btn btn-outline-primary rounded-pill px-4 py-2">🔐 Entrar</a>
             </div>
         </div>
     {% endif %}
