+++ templates/reset_password_request.html
@@ -20,7 +20,7 @@
             </div>
 
             <div class="text-center mt-3">
-                <a href="{{ url_for('login') }}" class="text-decoration-none">🔙 Voltar para o login</a>
+                <a href="{{ url_for('auth.login') }}" class="text-decoration-none">🔙 Voltar para o login</a>
             </div>
         </form>
     </div>
